Improving Log

+++ src/main/java/extractor/Main.java
@@ -9,6 +9,7 @@ import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
+import org.apache.commons.io.FileUtils;
 import org.limeprotocol.serialization.JacksonEnvelopeSerializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,6 +18,7 @@ import org.springframework.web.client.RestTemplate;
 import constant.ExtratorConstants;
 import controller.KBTLoadController;
 import enums.StatusFileEnum;
+import exception.ImportExcelException;
 import model.KnowledgeBase;
 import service.HttpService;
 import setting.KBTSettings;
@@ -39,9 +41,11 @@ public class Main {
 			return;
 		}
 		
-		Path filePath = initProcessKBT();
+		Path filePath = null;
 		
 		try {
+			filePath = initProcessKBT();
+			
 			//Extrai os dados da planilha
 			ExcelExtractor excelExtractor = new ExcelExtractor();
 			KnowledgeBase kb = excelExtractor.extractExcelData(filePath.toFile());
@@ -74,8 +78,9 @@ public class Main {
 	
 	/**
 	 * Move arquivo para a pasta Processando e loga início do processo
+	 * @throws ImportExcelException 
 	 */
-	private static Path initProcessKBT() {
+	private static Path initProcessKBT() throws ImportExcelException {
 		DateFormat df = new SimpleDateFormat("yyyyMMdd_hhmm");
 		String fileRename = StatusFileEnum.PROCESSING.getPathFile() + "kb_" + df.format(new Date()) + ".xlsx";
 		
@@ -83,11 +88,18 @@ public class Main {
 		Path targetProcessing = FileSystems.getDefault().getPath(fileRename);
 		
         try {
-            Files.move(movefrom, targetProcessing, StandardCopyOption.REPLACE_EXISTING);
-            LOGGER.info("Início do processo de atualização da base de conhecimento");
-            LOGGER.info("Arquivo movido para a pasta 'Processando.");
+        	if(Files.exists(movefrom)) {
+        		FileUtils.touch(movefrom.toFile());
+                Files.move(movefrom, targetProcessing, StandardCopyOption.REPLACE_EXISTING);
+                LOGGER.info("Início do processo de atualização da base de conhecimento");
+                LOGGER.info("Arquivo movido para a pasta 'Processando.");
+        	}
+        	else {
+        		throw new ImportExcelException("ERRO: O arquivo não está na pasta para ser processado");
+        	}
+        	
         } catch (IOException e) {
-            System.err.println(e);
+            throw new ImportExcelException("Não foi possível mover o arquivo para pasta processando. Favor verificar se o mesmo está aberto.");
         }
         
         return targetProcessing;
@@ -97,15 +109,17 @@ public class Main {
 	 * @param pathFile
 	 */
 	private static void moveFileToRejected(Path pathFile) {
-		String fileRename = StatusFileEnum.REJECTED.getPathFile() + pathFile.getFileName().toString();
-		Path pathToMove = FileSystems.getDefault().getPath(fileRename );
-		
-        try {
-            Files.move(pathFile, pathToMove, StandardCopyOption.REPLACE_EXISTING);
-            LOGGER.info("Falha ao processar arquivo. Arquivo movido para a pasta 'Rejeitados'");
-        } catch (IOException e) {
-            System.err.println(e);
-        }
+		if(pathFile != null) {
+			String fileRename = StatusFileEnum.REJECTED.getPathFile() + pathFile.getFileName().toString();
+			Path pathToMove = FileSystems.getDefault().getPath(fileRename );
+			
+	        try {
+	            Files.move(pathFile, pathToMove, StandardCopyOption.REPLACE_EXISTING);
+	            LOGGER.info("Falha ao processar arquivo. Movendo arquivo para a pasta 'Rejeitados'");
+	        } catch (IOException e) {
+	        	LOGGER.info("Não foi possível mover arquivo para a pasta 'Rejeitados'");
+	        }
+		}
 	}
